# swag/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API для сервиса mpit2026-reg.\n\nОсновные возможности:\n- авторизация через MAX mini app и локальный JWT\n- получение и обновление профиля текущего пользователя\n- создание событий, получение invite token и вход по нему\n- управление гостями и фиксация статуса посещения\n- wishlist и фото внутри события\n\nАутентификация:\n- защищённые методы требуют заголовок Authorization: Bearer \u003cjwt\u003e\n- JWT можно получить через /auth/max/login или через мобильный flow /auth/max/start -\u003e /auth/max/complete -\u003e /auth/max/exchange\n\nSwagger UI доступен по адресу /api/docs/.",
        "title": "mpit2026-reg API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "mpit-bot.kostya1024.ru",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/events/{eventID}/wishlist": {
            "get": {
                "description": "Возвращает список wishlist-позиций, привязанных к событию.\nСодержимое ответа может зависеть от роли пользователя: организатор видит больше деталей, гостям часть информации может быть скрыта.\nМетод используется для отображения списка подарков и текущего состояния бронирования или финансирования.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Получить wishlist события",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Массив wishlist-позиций события",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка при получении wishlist",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{eventID}/wishlist/ideas": {
            "post": {
                "description": "Принимает текстовое предложение гостя и проверяет его на соответствие anti-wishlist ограничениям.\nВ ответе backend возвращает, разрешена ли идея, а если нет — причину блокировки.\nЕсли идея допустима, в ответ также может быть возвращён созданный или подобранный item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Отправить идею подарка от гостя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса с текстом идеи подарка",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/action.SubmitGuestIdeaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат проверки идеи: allowed, reason, item",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID, JSON или пустой текст",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка при обработке идеи",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{eventID}/wishlist/parse": {
            "post": {
                "description": "Принимает произвольный текст и с помощью AI разбирает его на список желаемых подарков и anti-wishlist элементов.\nОбычно используется организатором события для первичного наполнения wishlist по заметкам, сообщениям или черновому описанию.\nВ ответе возвращаются массивы items и anti_items, сформированные на основе текста.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Разобрать свободный текст в wishlist и anti-wishlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса со свободным текстом для разбора",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/action.ParseWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сформированные массивы items и anti_items",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID, JSON или пустой текст",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка при AI-разборе wishlist",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{eventID}/wishlist/{itemID}/book": {
            "post": {
                "description": "Позволяет пользователю забронировать конкретный подарок полностью.\nПосле бронирования item нельзя частично финансировать или повторно бронировать другим участникам.\nЕсли позиция уже занята или по ней уже есть сбор средств, backend вернёт конфликт.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Забронировать wishlist item целиком",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор wishlist item",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном бронировании",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID или itemID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Item уже забронирован или по нему уже есть финансирование",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка при бронировании item",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{eventID}/wishlist/{itemID}/fund": {
            "post": {
                "description": "Позволяет пользователю внести сумму в общий сбор средств на конкретный подарок.\nИспользуется для сценария коллективного подарка, когда несколько гостей складываются на одну позицию.\nЕсли item уже полностью забронирован, backend вернёт конфликт.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlist"
                ],
                "summary": "Частично профинансировать wishlist item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор wishlist item",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса с суммой пополнения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/action.FundItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Текущее накопленное значение current_fund",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID, itemID, JSON или сумма",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Item уже полностью забронирован",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка при пополнении сбора",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "action.FundItemRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
...
        "action.ParseWishlistRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "action.SubmitGuestIdeaRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}